# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The Health Check is successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {},
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "The Health Check is not successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {
                        "redis": {
                          "status": "down",
                          "message": "Could not connect"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        },
                        "redis": {
                          "status": "down",
                          "message": "Could not connect"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Verificar el estado de las bases de datos",
        "tags": [
          "Health"
        ]
      }
    },
    "/users/search": {
      "get": {
        "operationId": "UsersController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "Término de búsqueda",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Límite de resultados (por defecto 20)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Buscar usuarios",
        "tags": [
          "Users"
        ]
      }
    },
    "/auth/spotify/login": {
      "get": {
        "operationId": "AuthController_loginSpotify",
        "parameters": [],
        "responses": {
          "302": {
            "description": "Redirige a la página de autorización de Spotify."
          }
        },
        "summary": "Iniciar login con Spotify",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/spotify/callback": {
      "get": {
        "operationId": "AuthController_callbackSpotify",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirige al frontend con el token en la URL."
          }
        },
        "summary": "Callback de Spotify Auth",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Perfil de usuario retornado."
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Obtener perfil del usuario autenticado",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sesión cerrada exitosamente."
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Cerrar sesión",
        "tags": [
          "Auth"
        ]
      }
    },
    "/spotify/now-playing": {
      "get": {
        "operationId": "SpotifyController_getNowPlaying",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Obtener la pista actual de Spotify",
        "tags": [
          "Spotify"
        ]
      }
    },
    "/spotify/recent": {
      "get": {
        "operationId": "SpotifyController_getRecent",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Límite de pistas (por defecto 20)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Obtener las pistas reproducidas recientemente",
        "tags": [
          "Spotify"
        ]
      }
    },
    "/friends": {
      "get": {
        "operationId": "FriendsController_getFriends",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Obtener lista de amigos",
        "tags": [
          "Friends"
        ]
      }
    },
    "/friends/requests": {
      "get": {
        "operationId": "FriendsController_getRequests",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Obtener solicitudes de amistad pendientes",
        "tags": [
          "Friends"
        ]
      },
      "post": {
        "operationId": "FriendsController_sendRequest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "targetUserId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Enviar solicitud de amistad",
        "tags": [
          "Friends"
        ]
      }
    },
    "/friends/requests/{id}": {
      "patch": {
        "operationId": "FriendsController_respondToRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de la solicitud",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "ACCEPT",
                      "REJECT"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Responder a una solicitud de amistad",
        "tags": [
          "Friends"
        ]
      }
    },
    "/playback/sync": {
      "post": {
        "operationId": "PlaybackController_sync",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "trackId": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "artist": {
                    "type": "string"
                  },
                  "albumName": {
                    "type": "string"
                  },
                  "albumImageUrl": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Sincronizar evento de reproducción",
        "tags": [
          "Playback"
        ]
      }
    },
    "/reactions": {
      "post": {
        "operationId": "ReactionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "playbackEventId": {
                    "type": "string"
                  },
                  "emoji": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Crear una reacción a un evento de reproducción",
        "tags": [
          "Reactions"
        ]
      }
    },
    "/reactions/{playbackEventId}": {
      "get": {
        "operationId": "ReactionsController_getByEvent",
        "parameters": [
          {
            "name": "playbackEventId",
            "required": true,
            "in": "path",
            "description": "ID del evento de reproducción",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Obtener reacciones por ID de evento de reproducción",
        "tags": [
          "Reactions"
        ]
      }
    },
    "/notes": {
      "post": {
        "operationId": "NotesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "playbackEventId": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Crear una nueva nota",
        "tags": [
          "Notes"
        ]
      }
    },
    "/notes/{playbackEventId}": {
      "get": {
        "operationId": "NotesController_getByEvent",
        "parameters": [
          {
            "name": "playbackEventId",
            "required": true,
            "in": "path",
            "description": "ID del evento de reproducción",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Obtener notas por ID de evento de reproducción",
        "tags": [
          "Notes"
        ]
      }
    },
    "/feed": {
      "get": {
        "operationId": "FeedController_getFeed",
        "parameters": [
          {
            "name": "friendId",
            "required": false,
            "in": "query",
            "description": "ID del amigo para filtrar",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Límite de resultados (por defecto 20)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Obtener el feed de actividad",
        "tags": [
          "Feed"
        ]
      }
    },
    "/conversations": {
      "get": {
        "operationId": "ConversationsController_getConversations",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Obtener conversaciones del usuario",
        "tags": [
          "Conversations"
        ]
      },
      "post": {
        "operationId": "ConversationsController_createConversation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "targetUserId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Crear una nueva conversación",
        "tags": [
          "Conversations"
        ]
      }
    },
    "/conversations/{id}/messages": {
      "get": {
        "operationId": "MessagesController_getMessages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de la conversación",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Límite de mensajes",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Mensajes a saltar",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Obtener mensajes de una conversación",
        "tags": [
          "Messages"
        ]
      }
    },
    "/messages/read": {
      "post": {
        "operationId": "MessagesController_markRead",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "conversationId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Marcar mensajes como leídos",
        "tags": [
          "Messages"
        ]
      }
    }
  },
  "info": {
    "title": "MusicTwins API",
    "description": "The MusicTwins API documentation",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt"
      }
    },
    "schemas": {}
  }
}